Rewrite media stream resolvers to be device profile based
