+++ playback/jellyfin/src/main/kotlin/mediastream/JellyfinMediaStreamResolver.kt
@@ -1,0 +1,117 @@
+package org.jellyfin.playback.jellyfin.mediastream
+
+import org.jellyfin.playback.core.mediastream.MediaConversionMethod
+import org.jellyfin.playback.core.mediastream.MediaStream
+import org.jellyfin.playback.core.mediastream.MediaStreamResolver
+import org.jellyfin.playback.core.mediastream.PlayableMediaStream
+import org.jellyfin.playback.core.queue.QueueEntry
+import org.jellyfin.playback.core.support.PlaySupportReport
+import org.jellyfin.playback.jellyfin.queue.baseItem
+import org.jellyfin.playback.jellyfin.queue.mediaSourceId
+import org.jellyfin.sdk.api.client.ApiClient
+import org.jellyfin.sdk.api.client.extensions.mediaInfoApi
+import org.jellyfin.sdk.api.client.extensions.videosApi
+import org.jellyfin.sdk.model.api.BaseItemDto
+import org.jellyfin.sdk.model.api.DeviceProfile
+import org.jellyfin.sdk.model.api.MediaProtocol
+import org.jellyfin.sdk.model.api.MediaType
+import org.jellyfin.sdk.model.api.PlaybackInfoDto
+
+class JellyfinMediaStreamResolver(
+	private val api: ApiClient,
+	private val deviceProfileBuilder: () -> DeviceProfile,
+) : MediaStreamResolver {
+	companion object {
+		private val supportedMediaTypes = arrayOf(MediaType.VIDEO, MediaType.AUDIO)
+	}
+
+	override suspend fun getStream(queueEntry: QueueEntry, testStream: (stream: MediaStream) -> PlaySupportReport): PlayableMediaStream? {
+		val baseItem = queueEntry.baseItem
+		if (baseItem == null || !supportedMediaTypes.contains(baseItem.mediaType)) return null
+
+		val mediaInfo = getPlaybackInfo(baseItem, queueEntry.mediaSourceId)
+
+		return when {
+			// Direct play
+			mediaInfo.mediaSource.supportsDirectPlay -> mediaInfo.toStream(
+				queueEntry = queueEntry,
+				conversionMethod = MediaConversionMethod.None,
+				url = api.videosApi.getVideoStreamUrl(
+					itemId = baseItem.id,
+					mediaSourceId = mediaInfo.mediaSource.id,
+					static = true,
+					tag = mediaInfo.mediaSource.eTag,
+				)
+			)
+
+			// Remux (direct stream)
+			mediaInfo.mediaSource.supportsDirectStream && mediaInfo.mediaSource.transcodingUrl != null -> mediaInfo.toStream(
+				queueEntry = queueEntry,
+				conversionMethod = MediaConversionMethod.Remux,
+				url = api.createUrl(requireNotNull(mediaInfo.mediaSource.transcodingUrl), ignorePathParameters = true)
+			)
+
+			// Transcode
+			mediaInfo.mediaSource.supportsTranscoding && mediaInfo.mediaSource.transcodingUrl != null -> mediaInfo.toStream(
+				queueEntry = queueEntry,
+				conversionMethod = MediaConversionMethod.Transcode,
+				url = api.createUrl(requireNotNull(mediaInfo.mediaSource.transcodingUrl), ignorePathParameters = true)
+			)
+
+			// No compatible stream found
+			else -> null
+		}
+	}
+
+	private suspend fun getPlaybackInfo(
+		item: BaseItemDto,
+		mediaSourceId: String? = null,
+	): MediaInfo {
+		val profile = deviceProfileBuilder()
+		val response by api.mediaInfoApi.getPostedPlaybackInfo(
+			itemId = item.id,
+			data = PlaybackInfoDto(
+				mediaSourceId = mediaSourceId,
+				deviceProfile = profile,
+				enableDirectPlay = true,
+				enableDirectStream = true,
+				enableTranscoding = true,
+				allowVideoStreamCopy = true,
+				allowAudioStreamCopy = true,
+				autoOpenLiveStream = false,
+			)
+		)
+
+		if (response.errorCode != null) {
+			error("Failed to get media info for item ${item.id} source ${mediaSourceId}: ${response.errorCode}")
+		}
+
+		val mediaSource = response.mediaSources
+			// Filter out invalid streams (like strm files)
+			.filter { it.protocol == MediaProtocol.FILE && !it.isRemote }
+			// Select first media source
+			.firstOrNull { mediaSourceId == null || it.id == mediaSourceId }
+
+		requireNotNull(mediaSource) {
+			"Failed to get media info for item ${item.id} source ${mediaSourceId}: media source missing in response"
+		}
+
+		return MediaInfo(
+			playSessionId = response.playSessionId.orEmpty(),
+			mediaSource = mediaSource
+		)
+	}
+
+	private fun MediaInfo.toStream(
+		queueEntry: QueueEntry,
+		conversionMethod: MediaConversionMethod,
+		url: String,
+	) = PlayableMediaStream(
+		identifier = playSessionId,
+		conversionMethod = conversionMethod,
+		container = getMediaStreamContainer(),
+		tracks = getTracks(),
+		queueEntry = queueEntry,
+		url = url,
+	)
+}
