+++ playback/jellyfin/src/main/kotlin/mediastream/MediaInfo.kt
@@ -1,0 +1,8 @@
+package org.jellyfin.playback.jellyfin.mediastream
+
+import org.jellyfin.sdk.model.api.MediaSourceInfo
+
+data class MediaInfo(
+	val playSessionId: String,
+	val mediaSource: MediaSourceInfo,
+)
